feat: allow user to specify shutdown signal
